+++ react/modules/pixelHelper.ts
@@ -8,6 +8,10 @@ export function mapCartItemToPixel(item: CartItem): PixelCartItem {
     productRefId: item.productRefId,
     brand: item.additionalInfo ? item.additionalInfo.brandName : '',
     category: productCategory(item),
+    detailUrl: item.detailUrl,
+    imageUrl: item.imageUrls
+      ? fixUrlProtocol(item.imageUrls.at3x)
+      : item.imageUrl || '',
   }
 }
 
@@ -25,7 +29,21 @@ export function mapBuyButtonItemToPixel(item: BuyButtonItem): PixelCartItem {
     productRefId: item.productRefId,
     brand: item.brand,
     category,
+    detailUrl: item.detailUrl,
+    imageUrl: item.imageUrl,
   }
+}
+
+/**
+ * URL comes like "//storecomponents.vteximg.com.br/arquivos/ids/155491"
+ * this function guarantees it comes with protocol in it.
+ */
+function fixUrlProtocol(url: string) {
+  if (!url || url.indexOf('http') === 0) {
+    return url
+  }
+
+  return 'https:' + url
 }
 
 /**
@@ -63,6 +81,8 @@ interface PixelCartItem {
   productRefId: string
   brand: string
   category: string
+  detailUrl: string
+  imageUrl: string
 }
 
 interface BuyButtonItem {
@@ -74,6 +94,8 @@ interface BuyButtonItem {
   productRefId: string
   brand: string
   category: string
+  detailUrl: string
+  imageUrl: string
 }
 
 interface CartItem {
@@ -88,4 +110,13 @@ interface CartItem {
   }
   productCategoryIds: string
   productCategories: Record<string, string>
+  detailUrl: string
+  // Field from the usual orderForm API
+  imageUrl?: string
+  // Field from the order-manager orderForm API
+  imageUrls?: {
+    at1x: string
+    at2x: string
+    at3x: string
+  }
 }